+++ src/main/kotlin/top/foxball/receptionbackendsystem/datasource/jdbc/CarRepository.kt
@@ -7,10 +7,10 @@ interface CarRepository : ReceptionRepository<Car, Int> {
     /**
      * 查询指定活动下的全部车辆。
      */
-    fun findByActivityId(activityId: Int): List<Car>
+    fun findByActivityId(activityId: Long): List<Car>
 
     /**
      * 查询指定活动下的指定车号车辆。
      */
-    fun findByActivityIdAndCarNumber(activityId: Int, carNumber: Long): Car?
+    fun findByActivityIdAndCarNumber(activityId: Long, carNumber: Long): Car?
 }
