+++ src/main/kotlin/top/foxball/receptionbackendsystem/service/ColorTagService.kt
@@ -3,7 +3,7 @@ package top.foxball.receptionbackendsystem.service
 import top.foxball.receptionbackendsystem.datasource.jdbc.ColorTag
 
 interface ColorTagService : ActivityEntityService<ColorTag, Int> {
-    fun findByActivityIdAndName(activityId: Int, name: String): ColorTag?
+    fun findByActivityIdAndName(activityId: Long, name: String): ColorTag?
 
-    fun existsByActivityIdAndName(activityId: Int, name: String): Boolean
+    fun existsByActivityIdAndName(activityId: Long, name: String): Boolean
 }
