+++ src/main/kotlin/top/foxball/receptionbackendsystem/service/PersonService.kt
@@ -5,5 +5,5 @@ import top.foxball.receptionbackendsystem.datasource.jdbc.Person
 interface PersonService : ActivityEntityService<Person, Int> {
     fun findByNameContaining(name: String): List<Person>
 
-    fun findByActivityIdAndUnit(activityId: Int, unit: String): List<Person>
+    fun findByActivityIdAndUnit(activityId: Long, unit: String): List<Person>
 }
